use the objects in the client and vice versa
Objects now have few methods to interact with the db via the client
(typically, create, save, delete).
On the other end, the client will return objects when possible.

Objects can also be used as input for several methods, but no
modification of the interface is needed: on should just call the
getMap() method and use the ** operator to develop the dictionnary. The
only drawback is that it fails when additional members are defined (rare).