implemented the last object unit tests
All the tests so far are patching requests not to use any server. The
client is not tested in a detailed way. Only methods called by objects
are checked against expected server responses.

Among the objects, higj-level tree-manipulation classes are not tested
either, as this requires a real server to make sense.

Next: validate the client with a test server